pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.2.0"
description = ""
authors = ["GrassSand <grasssand@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.2.2"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"
aiohttp = "<3.8.0"
aiofiles = "^0.7.0"
aioredis = "^2.0.0"
SQLAlchemy = "^1.4.27"
Pillow = "^8.4.0"
genshin = "^0.1.0"
cachetools = "^4.2.4"

[tool.poetry.dev-dependencies]
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
